GtkTreeView Part3: replace GtkCTree in Bookmarks dialog (bookmarks.c)
The same behavior and functionality is retained but the TreeView
is set to not be reorderable, the GtkCTree was reorderable and supported drag'n'drop.
It's simpler but the code is more complex.

Making the TreeView reorderable will complicate things to an unknown degree.

The bookmarks hash table and the GtkCTree used to have
references to each other, but now only the TreeView has pointers
to bookmark entries.

Bookmark->cnode was a pointer to a GtkCTreeNode.
Now it's just a flag variable (or a hack).