src/gtk: Replace calls of gtk_window_set_wmclass to gtk_window_set_role
According to the documentation:

Don't use this function. It sets the X Window System "class" and "name" hints for a window. According to the ICCCM, you should always set these to the same value for all windows in an application, and GTK+ sets them to that value by default, so calling this function is sort of pointless. However, you may want to call gtk_window_set_role() on each window in your application, for the benefit of the session manager. Setting the role allows the window manager to restore window positions when loading a saved session.

This has the side-effect of making the application more compatible with non-Xlib based platforms